chore: Add LICENSE file
The LICENSE file is added to the project to clarify the permissions and
conditions for using, copying, modifying, merging, publishing, distributing,
sublicensing, and selling the software. The copyright notice and permission
notice are included in all copies or substantial portions of the Software.
The software is provided "as is" without warranty. The addition of the LICENSE
file ensures proper documentation and protection for the project.