feat: Add Pest testing framework to project
- Add Pest as a development requirement with version ^2.33
- Add configuration to allow Pest plugin
- Add a script for running tests using Pest
- Remove unused "require" key in composer.json

The Pest testing framework was added to the project as a development
requirement. This will enable us to write and execute tests for our
codebase more easily. A configuration was also added to allow the use of
the Pest plugin. Additionally, a new script was created to run tests
using Pest. Finally, the unused "require" key in composer.json was
removed to keep the file clean and clear.