feat: Add RouterTest for testing routes execution
This commit adds the RouterTest class for testing the execution of
routes in the Hermes Router package. The RouterTest.php file has been
created with three test cases.

The first test case checks the execution of a GET route. The route is
defined using the `Router::get()` method, which sets the callback
function to echo "GET route". The test then sets the `$_SERVER`
variables to simulate a GET request to '/test' and captures the output
using output buffering. Finally, the test asserts that the output
matches 'GET route'.

The second test case checks the execution of a POST route. Similar to
the first test case, the route is defined using the `Router::post()`
method with a callback function that echoes 'POST Test'. The test
simulates a POST request to '/test', captures the output, and asserts
that the output matches 'POST Test'.

The third test case checks the execution of a route that supports
multiple HTTP methods. The route is defined using the `Router::route()`
method with an array of HTTP methods, 'GET' and 'POST', and a callback
function that echoes 'GET and POST Test'. The test first simulates a
POST request to '/test', captures the output, and asserts that the
output matches 'GET and POST Test'. Then, it simulates a GET request to
'/test', captures the output, and asserts that the output matches 'GET
and POST Test'.

These test cases ensure that the Router class can properly execute and
handle different types of routes in the Hermes package.