feat: Add support for additional HTTP methods
This commit adds support for the HTTP methods `PUT`, `DELETE`, `HEAD`, `OPTIONS`, and `PATCH` in the `Router` class. These methods can now be used to define routes and associate actions with them. This increases flexibility in handling different types of requests and provides more options for developers when designing their routing logic. With this update, the `Router` class is now more robust and capable of handling a wider range of HTTP methods.