fix: Tests script in composer.json
- Removed "tests" script value "./vendor/bin/pest tests"
- Updated "tests" script value to "./vendor/bin/pest"

Reason for the change:
The previous value for the "tests" script in composer.json was incorrect. It included the argument "tests", which is not necessary. Therefore, the script was updated to the correct value of "./vendor/bin/pest". This change ensures that the correct command is executed when running tests.