fix(Router): Set proper HTTP response code for error routes
The Router class has been modified to set the correct HTTP response code for the error routes.

In the initialize() method, the HTTP response code for the "404" route has been set to 404. This ensures that when the page is not found, the server returns the appropriate HTTP response code.

Additionally, for the "500" route, the HTTP response code has been set to 500. This will be helpful in identifying internal server errors and informing users about them.

The changes were made to improve the error handling functionality of the Router class and provide meaningful feedback to users in case of errors.