chore: Add keywords and homepage to composer.json
Adding keywords and a homepage to the composer.json file will provide more information and visibility for the Hermes library. The keywords "routing", "router", and "web" will help users find the library when searching for related functionality. The homepage link to the GitHub repository will allow users to easily access and contribute to the project. This will improve the overall usability and discoverability of the Hermes library.