feat: Add support for multiple HTTP methods in route definition
This commit adds support for defining routes with multiple HTTP methods
in the Hermes library. Previously, only GET and POST methods were
supported, but now it is possible to define routes with GET, POST, PUT,
DELETE, HEAD, OPTIONS, and PATCH methods.

This feature was added to enhance the flexibility of the library and
provide developers with more options when designing their routing
logic. Being able to define routes with different methods allows for more
granular control over how requests are handled and processed.

In addition to the new method support, this commit also includes an
example in the README.md file showcasing how to define a route that
executes for two or more HTTP methods. The example demonstrates the use
of the `Router::route()` static method and provides a code snippet to
illustrate the implementation.

Overall, this feature enhances the usability and versatility of the
Hermes library, making it easier for developers to create robust routing
systems in their PHP projects.